reorder and rename files
