+++ lib/twitter_api/twitter_api.dart
@@ -4,9 +4,8 @@ import 'package:cest_pret_de_chez_vous/models/ad.dart';
 import 'package:cest_pret_de_chez_vous/models/category.dart';
 import 'package:oauth1/oauth1.dart';
 
-part './tweet_from_ad.dart';
-part './tweet_to_ad.dart';
-part './twitter_api_client_utils.dart';
+part './post_tweet_utils.dart';
+part './fetch_tweets_utils.dart';
 part './twitter.config.dart';
 
 part 'twitter_api_client.dart';
